feat: refactor setCurrentLocation to util folder
